NEED TO CHANGE UP DATABASE
the table for EOM_CLASS need to be changed up so CLASS and YEAR are combined into one thing. I am going to have to use string sorting thing to break up the one string into course code and year. ICS4U-01 2018 will be the class name from now on.